Merged PR 709500: Change the error logging message displayed on ADO console for DX64 errors.
Adding changes to AzureDevOpsListener.cs to change the error logging message displayed on ADO for DX64 errors.
Modified the unit tests to accommodate the corresponding coding changes.

Tested the changes in Primary PR validations to check if it meets the requirements.
Before:
https://dev.azure.com/mseng/Domino/_build/results?buildId=19569483&view=logs&j=275f1d19-1bd8-5591-b06b-07d489ea915a&t=1514d7bc-8bce-462c-8ffb-d9a2e471b2b3&l=46

After:
https://dev.azure.com/mseng/Domino/_build/results?buildId=19599846&view=logs&j=275f1d19-1bd8-5591-b06b-07d489ea915a&t=f63fd35e-f37e-45df-8a53-c1633fd2ed1a&l=52

This change highlights only OutputToLog segment of DX64 error message as we want only this part to be highlighted.

Attaching another PR which has another modified version of the same code to refer for comments.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/709377

From this PR highlighting the main concerns
1. Refactored the LogDevOpsIssue code.
2. Created another Specific story for PipProcessWarnings.
3. Addresed other comments like Renaming of the ADOFormatting method, removal of New lines.
4. Modified MockConsole.cs for unit tests to check the highlighting changes.
5. Replaced array usage for event payload with string interpolation.