Merged PR 710877: Avoid traversing the same augmented weak fingerprint candidates more than once
A multi-layer cache can result in the same (wf -> augmented wf) pair returned more than once. Today this means we will go through all the candidates associated with the augmented wf as many times as duplicate pairs come back. This PR shortcuts that search when it sees the same augmented wf coming again for the same lookup.

Sorry, no tests for this optimization. The only way to get duplicate entries is by having an at least 2-layer cache, which is pretty complicated with the test infra we have today.