Merged PR 711267: Merge Interop into Utilities.Core
Merge Interop to Utilities.Core. The recommended way to create Nuget packages is to do 1 dll per package unless the dlls are supposed to always be used together. Interop is a part of Utilities and Utilities.Core depends on it meaning that it's already an indirect dependency to basically everything, so merging them together to reduce the amount of dlls.