Merged PR 711426: Detect and warn when CacheDump fingerprint computation is not accurate
The WeakFingerprint text displayed by CacheDump is based on recomputing the fingerprint based on data reconstructed by the recorded XLG events. We've discovered cases where this reconstruction does not match what was calculated when the build was actually running.

The cause of this isn't totally understood at this time. A cause may be that the fingerprint reconstruction is based on the accumulated filesystem state at the end of the build rather than at the timestamp within the build when the fingerprint was computed. That doesn't seem to be the case in the build referenced in the related bug.

This change doesn't fix the source of the inconsistency. It detects it and warns the user when the fingerprint text doesn't match.

Related work items: #2033249