Merged PR 712282: Change the warning logging message displayed on ADO console for DX65 warnings.
Created a new PR as the old one has a lot of test commits.

Adding changes to AzureDevOpsListener.cs to change the warning logging message displayed on ADO for DX65 warnings.
Modified the unit tests to accommodate the corresponding coding changes in DevOpsListener.cs
Modified Interfaces.Proto file accordingly to ensure the PipProcessEvent works for PipProcessWarnings and PipProcessErrors.
Refactored PipProcessErrorEventFields class to accommodate both PipProcessErrorEvent and PipProcessWarningEvents.
Created another constructor for PipProcessWarning event to avoid issues related to default values.

Before:
https://dev.azure.com/mseng/Domino/_build/results?buildId=19672141&view=logs&j=275f1d19-1bd8-5591-b06b-07d489ea915a&t=1e890290-1da7-4ac9-b698-21f5ae7ff179

After:
https://dev.azure.com/mseng/Domino/_build/results?buildId=19671509&view=logs&j=275f1d19-1bd8-5591-b06b-07d489ea915a&t=9dd78ff8-514f-446d-b196-fc754a84acf9&l=375

This change highlights only OutputToLog segment of DX65 warning message as we want only this part to be highlighted.

Adding the previous PR as that has some of the comments related to this PR.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/711253