Merged PR 713148: Fix crash on CB timeout shutdown
Timeouts in CloudBuild are crashing because an error is not logged before the cancellation signal is sent. This seems to have been caused by !696346, which was trying to address some logging oddities when an early termination cancellation causes an internal error to be logged which then triggers the internal error shutdown.

This change reverts the switch made in that previous PR and makes the internal error early termination ignore timeout cancellations. This should address the crash as well as the funny logging on timeout.

Related work items: #2042492, #2044279, #2044284