Merged PR 714149: Force PTrace sandbox for specific process names
Allow the user to provide a semicolon-separated list of processes that will be forced to run under the PTrace sandbox rather than the LD_PRELOAD sandbox.

This is a debugging measure so I didn't do the effort to add it into the FileAccessManifest, which would be more clean. The FAM serialization code is shared across Windows and Unix so increasing complexity there for this scenario is not warranted (unless we want to make it permanent)