Merged PR 714356: Fix for SaveFileChangeTrackerAsync: NullReferenceException bug
The cause for this bug is that **m_filechnageTracker** in SaveFileChangeTrackerAsync method of Scheduler.cs was not initialized when the build is cancelled.

https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/Scheduler/Scheduler.cs&version=GBmain&line=6460&lineEnd=6460&lineStartColumn=22&lineEndColumn=43&lineStyle=plain&_a=contents
This method initialized **m_fileChangeTracker** object.

If we follow the parent method references we see that in **InitForOrchestrator** method the parent method for InitFileChangeTracker is invoked.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/Scheduler/Scheduler.cs&version=GBmain&line=6064&lineEnd=6065&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents
However this wont be called if there is a failure.
The logs received from the customer state that the user has used ctrl + c. This caused a failure thereby not initalizing the object.
Added a fix to resolve this.

Related work items: #2042494