Merged PR 715153: Revert source file tracking in some static mounts on Windows
Many X-plat tools get installed in user profile (or home folder in Linux), or in local app data (`$HOME/.local` in Linux), or in program data (`/usr/local/bin` or `/usr/bin` in Linux). We want to make sure that we can track and hash the files in these locations, so we add them as static mounts, and enable source file change tracking on them.

In general the enablement of source file change tracking is needed when the process pip allows for undeclared source reads. Otherwise, the observed input processor will see a read observation on a file that it cannot hash/track, and thus it will emits an error.

For now, the enablement of source file change tracking is done only on Linux, since enabling it on Windows has caused some performance issues in some customer builds. For example, a file in a user profile specified as a source file in the build suddenly gets tracked, and since this file keeps changing and is consumed by many pips, those pips get dirty all the time, and the builds slow down significantly due to losing incremental scheduling.

Related work items: #2060740