Merged PR 715943: Untrack $Extend\$Deleted from Windows sandbox observation
Windows can have an "unlink" behavior where deleted files are not really deleted if there's an opened handle. Note that this behavior is possible because a process can open a file with FILE_SHARE_DELETE that makes other process able to delete it.

See this unofficial doc: https://dfir.ru/2020/03/21/the-extenddeleted-directory/

Since this is an internal Windows behavior, such accesses need to be untracked.

We don't want users to explicitly specify the untracked scopes. Further, we want other sandbox's consumers, like QuickBuild and AnyBuild, to get this untracking behavior. So, the untracking behavior change is implemented down at the Detours layer.

Creating a unit test to explicitly writing to \$Extend\$Deleted is not possible due to access denied, even when the test is run in admin mode.

Related work items: #2061885