Merged PR 716305: AddOrGetContentHashList traces the hash count of the persisted CHL
AddOrGetContentHashList traces the hash count of the persisted CHL. This will be very useful when attempting to calculate a theoretical cost of a build with a ring-wise L2 and an Azure L3. In theory, once we have this, we would be able to roughly use the formulas:

Reads = count(GetCHL) + sum(GetCHL.HashCount)
Writes = count(AddCHL) + sum(AddCHL.HashCount)

This would give us a worst-case scenario cost for a given build with that particular cache configuration.