Merged PR 716623: Add missing dependency for Microsoft.Identity.Client
The nuget generation logic seems to have a problem with a particular dependency, and even though Microsoft.IdentityModel lists Microsoft.IdentityModel.Abstractions as a dependency, the generated nuget spec does not include the latter for net6 (it does for other frameworks).
Patching this manually by including the missing dependency explicitly. We'll file a bug for fixing the root cause.