Merged PR 717357: Consider monikers in nuspec dependency tfms
It seems that nuspec files can express nuget package dependencies for a given target framework not only by using the framework name, but also by using the moniker (e.g. ".NetCore6.0" vs "net6.0").

The fix makes nuget packages to have their proper dependency closure, but it generates a bunch of issues in our selfhost that can't be fixed without this change. Therefore, this is done under a temporary flag until we can deploy the change.

In addition, another bug is fixed (rolled out under the same flag) where NetCoreApp2.1 was wronly placed after net6.0, and other frameworks it is compatible with.

The flow would be: get this change into the LKG -> merge all the fixes in our selfhost + turning on the new flag -> get this change into the LKG -> flip the flag default (and optionally remove the flag altogether)

Related work items: #2062926