Merged PR 717827: Increase unit test concurrency limit when running in CB
Allow for a larger concurrency of unit tests in cloud build because the machine has more cores. Given that a unit test can be heavy, we limit the number of concurrent tests by making the unit test pips heavier (weight = 2).

With this change, the queue duration time in the critical path has reduced from 6-7mins to just 3-7s.

Related work items: #2042279