Merged PR 718929: Fix classification of rooted path in Linux build when parsing config
If config has a windows path, and it's evaluated on Linux, the evaluation fails.

When parsing configuration, we set `convertPathLikeLiteralsAtParseTime` to false, and so we are using the TypeScript parser. The path conversion will be done later by the AST converter.

For others, we set `convertPathLikeLiteralsAtParseTime` to true, and so we are using DScriptParser. DScriptParser is responsible for converting the path.

The path conversions in DScriptParser and AST converter have different logics for paths like C:/foo/bar.  In linux, the former one says that it's a relative path, but the latter says it should be an absolute path due to the colon `:` , but it's an invalid absolute path for linux. Thus the AST converter complains about such a path.