Merged PR 719100: Fix calculation for graph shape limiting resource
Currently for GraphShape indicator, we compute the limiting factor based on the number of ready pips. However, the transition from ready to running state is so quick that the number of ready pips is often 0.

In this PR, in addition to the number of ready pips, the calculation of GraphShape includes the number of running pips that have not been executed in the pip executor. I.e., we use the number of pending pips.

Related work items: #2066391