Merged PR 717989: [ADO distributed builds] Let the worker pipeline be triggered in a different job
Instead of communicating the build id and the orchestrator location at trigger-time, only pass along the build id. The AdoBuildRunner on the worker side will query the triggering build's `properties` (an ADO REST API endpoint), that the orchestrator run will use to publish the build information (namely, related session ID and the orchestrator location for gRPC). To access the build information we use a "invocation key" that will identify a build invocation in a particular pipeline run that might have many (e.g.: "Debug"/"Ship", "Metabuild"/"ProductBuild").

This allows the trigger of the worker pipeline to happen separately from the job that runs the build: with this, we can launch the worker pipeline from an agentless task at the beginning of the build, with the objective of improving the synchronization of the workers' provisioning.

Besides this, some additional changes:
- We now use the machine name as the location instead of the IP. This works, looks better (for logging too), and does not require the querying of the IPs to the network interfaces.
- Some "state consistency" sanity checks are added, to ensure the user has correct YAML specifications for the distributed build: namely, we verify that:
    1. the pipelines are running on the same branch and commit ID (so the initial sources should in principle be identical),
    1. only one (possibly parallel) job might act as "the workers" for a particular invocation key: this is to avoid copy-paste issues where the same invocation key ends up in multiple "worker job" specifications. This is done on a best-effort, racy manner, but the idea is that we catch this mistake so the user can fix it, not that the verification is bulletproof every time.

Related work items: #2066329, #2066330