Merged PR 719768: Add GenericGrpcClientPool for pooling instances that wrap gRPC.NET clients
This is required for upcoming PRs where I'll make use of this to allow having a wrapper class around a gRPC.NET client interface. This wasn't needed until now because GCS only has a single client so we'd just created that one client, but we'll be using a lot of these for content tracking and so forth.

I removed the try catch statement because it messes around with the retry policies we already have for timeouts.