Merged PR 719775: Add in-memory content tracking that observes timestamps
This PR adds a mechanism for in-memory content tracking that (obviously) doesn't depend on RocksDb. The mechanism is implemented on top of a LWW-Element-Set [CRDT](https://en.wikipedia.org/wiki/Conflict-free_replicated_data_type). There's no gossipping in this PR, it will be introduced in a latter one.

Please note, there's no GC in the content tracker basically on purpose: we don't expect to live too long :)