Merged PR 720794: Propagate a failure in ReportFileArtifactPlaced without marking the materialization task as a success
ReportFileArtifactPlaced can fail (on failure of the file operations carried out: TrySetExecutePermissionIfNeeded, UsingFileHandleAndFileLength) and return a `Failure`. Because we do this on "materialization success" (the actual materialization did succeed), we were marking the materialization task as completed with success. But returning a failure at this point causes the subsequent failure logic to kick in and mark the task as a failure, throwing an exception due to the task being completed, and ultimately crashing the engine.

Failure of these operations implies we didn't succeed in the "full" materialization, so marking it as a failure overall seems appropriate, we just need to skip the SetMaterializationSuccess.

Related work items: #2065572