Merged PR 722614: Dump /proc/self/stat upon failure collecting process information
We've observed failures to collect the information on Linux (see bug #2070047)
According to the dotnet team, this may happen if something happens reading/processing the process' stat file (see https://github.com/dotnet/runtime/issues/87356).

Let's try to get the file contents to hopefully help understand this issue better. We should remove that logic after getting a few examples. We've observed this to happen a couple of times per week so it shouldn't be long before getting an example.

Related work items: #2073411