Merged PR 722290: Add pin eliding capabilities based on a configured storage retention policy
Add pin eliding capabilities based on a configured storage retention policy. This allows users to configure a blob storage-backed L3 cache to use its off the shelf deletion policy based on last access time as a way to do garbage collection. When the configured retention time is provided to the cache configuration, it can be used to determine when it is safe to elide pins.