Merged PR 723716: Optionally, add symbol indexing data to BSI
This is a new version of the PR that add symbol data to bsi. There are several consumers of this data that need to be updated first. This pr is essentially the same PR that we had before, with 1) feature being off by default, and 2) an env var is used to enable it.

Related work items: #2060729