Merged PR 725421: Add the ability to selectively send logs to the console
The idea is that a new listener is created that selectively sends logs based on a user configurable knob to an internal Logger that the console listener listens to (in addition to the standard Events source). When the (only) event in sent to the internal logger, the console listener displays that in the console 'as is'.

Related work items: #2074267