Merged PR 725423: Remove ObservedPathSet from ExecutionResult
ObservedPathSet causes lots of overhead when we send it from workers to the orchestrator as a part of ExecutionResult. The main reason is the big number of uncompressed paths in the string form. We already send ObservedPathSet as a part of ProcessFingerprintComputationEventData in the execution log. In the latter form, the paths are optimized as those are being serialized as a part of ExecutionLog. Because we keep track of the paths we have sent, we try to send strings that are as compact as possible. Please see BinaryLogger -> Write(AbsolutePath).

We only need ObservedPathSet on the orchestrator to report it to HistoricMetadataCache. With my changes, we do not send ObservedPathSet along with ExecutionResult. We just consume the pathsets from ProcessFingerprintComputationEventData events in OchestratorSpecificExecutionLogTarget.

The files with important changes:
- OchestratorSpecificExecutionLogTarget.cs
- Scheduler.cs
- ExecutionResultSerializer.cs

Here is the benefit:
BEFORE -> 100% cache hit - 43 minutes  https://cloudbuild.microsoft.com/build/dbecdd35-57d6-45a1-9610-4d6e986a8ddc?bq=OFFICECLIENT_2302_DEBUG_DROID

AFTER -> 100% cache hit - 26 minutes https://cloudbuild.microsoft.com/build/1195ab2b-cd36-49d2-a588-877ad41037b3?bq=OfficeClient_2302_Debug_Droid

Related work items: #2064945