Merged PR 723790: Workflow SDK
The workflow SDK provides abstractions that simplify build logic in DScript. Instead of modelling the build logic as low-level processes and file/directory dependencies between them, the SDK allows for expressing the build logic as dependencies between high-level tasks and projects.

The SDK is introduced to lower the cognitive load when the user needs to write and maintain DScript specifications when onboarding to BuildXL.

How to review:
- Review the documentation in the Public\Sdk\Public\Workflow\Docs folder of the SDK.
- Review the unit tests at Public\Sdk\UnitTests\Workflow\Test.Workflow.dsc.
- Feel free to go into the details of the SDK.

Related work items: #2073478, #2073483