Merged PR 726534: Limit the amount of open connections from the blob storage cache by sharing an HttpClient between BlobClients
If left unbounded, we have observed spikes of >65k open sockets (at which point we hit the OS limit of open files for the process - on Linux, where sockets count as files). Running builds where we limit this value all the way down to 100 didn't see any noticeable performance impact, so 30k shouldn't pose a problem. The configurable limit is per-client and per-server, but because we will reuse this HttpClient for all BlobClients and the 'server' (blob storage endpoint) is also always the same, we are effectively limiting the number of open connections in general.

Related work items: #2076905