Merged PR 724776: BlobLifetimeManager in charge of L3 garbage collection.
Currently, there is no garbage collection for the L3 cache. The default garbage collection available through Azure storage is no good to us, since it would not guarantee that all fingerprints are fully backed (its content available for download), so we have to build our own garbage collection for the blob L3.

The idea here is that we keep a reference count for each piece of content, and we evict content hash lists (CHLs) in LRU ordering, only deleting content once its reference count reaches 0.

The most important pieces of code are in the `BlobLifetimeManager.Library` assembly. For now, there are essentially 3 big parts:
- `RocksDbLifetimeDatabase`: a RocksDB-backed database that stores information about blob sizes and content reference counts, with LRU enumeration capabilities. Special attention to the LRU enumeration code would be appreciated.
- `LifetimeDatabaseCreator` enumerates the blobs in the remote and constructs a brand-new database from it.
- `BlobLifetimeManager`: uses the database to perform garbage collection as previously described. It will make checks with the remote to make sure that its knowledge about last access times is accurate before performing any deletion.