Merged PR 726602: Update critical path summary
We log the breakdown of the critical path at the end of the build. It has become somewhat complicated, and the nested relationships are quite unclear, such as which element is under which. That's why I've tried to make it as clear as possible.

Here is before and after. As you notice, I do not log zero durations anymore. The breakdown of the scheduler duration, critical path and each step is very clear now.

**BEFORE**:
![before (2).png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/726602/attachments/before%20%282%29.png)

**AFTER**:
sample1
![after1.png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/726602/attachments/after1.png)

sample2
![sample1 (2).png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/726602/attachments/sample1%20%282%29.png)

sample3
![sample2 (2).png](https://mseng.visualstudio.com/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/726602/attachments/sample2%20%282%29.png)

Related work items: #2020444, #2064945, #2079624