Merged PR 729087: Separate cache miss types and their associated counters
Previously the values of cache miss types are aligned with their counters. This can cause an issue when performing analysis with BXL that has a different version from the on producing the execution log. The BXL used for performing the analysis may include changes that modify the counter enums. When doing the analysis, the analysis shows no error, but the miss type shown can be misleading because the cache miss types have different values.

Ideally, one uses the same BXL for analysis as the one used for producing the execution log. However, there's nothing preventing the customers from using a different version of BXL so long as the execution log format is compatible.

This PR also includes a fix in counting cache hits and misses.

Related work items: #2082529