Merged PR 729365: Fix for pathset explosion: Always include static declaration in directory membership with minimal graph
The pathset explosion is caused by different directory membership fingerprints during cache lookup and after execution.

1. In the first run, the pip executes, and the membership of directory X calculated using minimal graph is [a, b, c, d].
2. In the second run, during cache lookup, the membership calculated of directory X using minimal graph is [a, b, c, d, e]. The pip executes again, and after execution the membership of X using minimal graph is again [a, b, c, d]. However, since the file accesses are non-deterministic, the pip creates another pathset.
3. 1 & 2 repeats causing pathset explosion.

1 & 2 show that the pip hardly ever gets a cache hit.

The reason for this different memberships is because the file e declared as a source file is under a partial sealed directory, and e is non-existent.

During the cache lookup, the hash of e has not been registered to the file content manager, and so it will be included in the pip FS view, and thus will be included in membership. During the execution, the pip will 'read' e, and thus during the observed input processing, the hash of e (absent hash) will be registered to the file content manager. When calculating the directory membership fingerprint of X during the observe input processing (after execution), the hash of e is known to be absent, an so it is not added to the pip FS view, and thus not included in the membership.

Related work items: #2082529