Merged PR 729621: Allow users to limit the number of unique path sets to check during cache lookup
This is an opt-in feature that can be used to avoid checking too many path sets due to path set explosion.

One can pass `/limitPathSetsOnCacheLookup+` to limit the number of unique path sets to the default value.
One can pass `/limitPathSetsOnCacheLookup:N` to liimt the number of unique path sets to `X`.

When disabled, the limit is assumed to be `int.MaxValue`.

Related work items: #2085907