Merged PR 728817: [Detours] Avoid memory allocation when always performing remote injection (part 2)
If the detoured process is WoW64 and the manifest dictates that Detours injection is always remote, i.e., performed by the managed side, then on initializing the injector of that detoured process, we don't need to set the payload because it is never used.

This saves a lot of memory when detouring a 32-bit process.

Related work items: #2083082