Merged PR 729739: Add noticeFile to BuildXL nuget packages when exists
The notice file compiles the license and copyright information for any code or other materials under open source licenses that we distribute in a Microsoft Offering. It is automatically generated in our rolling builds before we execute the selfhost build that produce the nuget packages. In those rolling builds, the notice file is put on the source root.