Merged PR 731127: Use ActionBlockSlim as draining queue for binary logger
This PR replaces the use of dedicated thread for draining the events from the concurrent queue with ActionBlockSlim.

The results below show that we create much less event writers and much shorter pending events.

**Results**

*OfficeTenant_CleanNmakeFromRoot_Setup_Retail_Mularchy -- Before*
  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=154,406
  ExecutionLogFileTarget.MaxPendingEvents=2,118,602
  ExecutionLogFileTarget.EventWriterFactoryCalls=123,312,139

*OfficeTenant_CleanNmakeFromRoot_Setup_Retail_Mularchy -- After*
  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=67,271
  ExecutionLogFileTarget.MaxPendingEvents=835,368
  ExecutionLogFileTarget.EventWriterFactoryCalls=51,719,290

*BuildXL selfhost -- Before*
  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=662
  ExecutionLogFileTarget.MaxPendingEvents=14,679
  ExecutionLogFileTarget.EventWriterFactoryCalls=84,342

  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=473
  ExecutionLogFileTarget.MaxPendingEvents=16,595
  ExecutionLogFileTarget.EventWriterFactoryCalls=84,180

*BuildXL selfhost -- After*
  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=289
  ExecutionLogFileTarget.MaxPendingEvents=5,828
  ExecutionLogFileTarget.EventWriterFactoryCalls=22,450

  ExecutionLogFileTarget.ExecutionLogFileLoggingTimeMs=280
  ExecutionLogFileTarget.MaxPendingEvents=9,580
  ExecutionLogFileTarget.EventWriterFactoryCalls=26,754

**Context**
With the use of dedicated thread for draining the events from the concurrent queue, we see a very long pending events in the queue, which means that BuildXL holds a lot of memory just for writing events. Also, since there can be a massive amount of events and the pool is relatively small, instead of getting an event writer from the pool, BuildXL keeps creating event writers. This event writer is not a small object either, i.e., it includes an instance of memory stream and a stack in it.

The design of binary logger itself is unique, i.e., it allows for events to be logged in a multithreaded manner, but at the same time it must preserve some events that are meant to be serial. This serialization has to be maintained, otherwise some analyzers will stop working. Also, since all events are eventually be written to a file, then they have to be serialized.

To this end, the binary logger uses an event writer object to stage a single event writing to a memory stream. Then the event writer is queued for serial writing to the file stream.

For example,

Thread 1: Event A, Event B (some analyzer requires A to come before B)
Thread 2: Event C.

The end result in the file can be
- Event A, Event B, Event C
- Event C, Event A, Event B
- Event A, Event C, Event B

Throttling the event writings to enable more reuse of pooled event writers or to shorten the pending queue results in slowdown.