Merged PR 727020: Report all DFA's when there is a retry.
Modified the method RetryProcessDueToUserSpecifiedExitCode to ensure that we pass unexpectedFileAccess during the creation of SandboxedProcessPipExecutionResult object.

Consider ExecutePipAndHandleRetryAsync method in PipExecutor class. We will observe that the retry is inline(within the same worker). The SandboxedProcessPipExecutionResult  object is being not returned until all the inline retries have been completed. A new linkedlist has been created here to capture all the previous SandboxedProcessPipExecutionResult objects, which contains all the UnexpectedFileAccess and SharedDynamicDirectoryWriteAccess. We do a union of the required lists before we return the final ExecutionResult object.
Note: We do not perform a union AllowedUndeclaredSourceReads or DynamicObservations during an inline retry, as they are reported by ValidateObservedFileAccessesAsync and it is not desirable to call this method for every inline retry given the existing design.
 We may not be able to capture some of the above mentioned ObservedFileAccesses during an inline retry(they are captured during a reschedule retry), but the effect is minimal as it a read access.
Discussed this with @<Serge Mera> offline to ensure this does not break anything.
Also we ensure that only the latest SandboxedProcessPipExecutionResult is used for fingerprint computation among all the retries(which is what we were doing before as well). This is done to ensure that we do not store all the results from previous retries.

Consider ExecutePipStep method in the Scheduler class. After the pip execution is done and the ExecutionResult has been obtained. It is added to another list ExecutionResultsForRetry created in the ProcessRunnablePip object. Which is used for similar purpose as above to do a union of all the violations and allowlist before sending it for a final DFA check in post process.

During the post process we do a union of all the required allowlists, violations, write and read access, dynamic observations, from all the retries, before sending for a final DFA check.

Two unit tests to test the changes
One in BaselineTests.cs - ReportDFAWhenRetry which is detour based.
The other one is in SharedOpaqueDirectoryTests - TwoPipsProducingTheSameFileDynamicallyUnderASharedOpaqueDirectoryIsBlockedEvenOnARetry which checks for double write violations.

Related work items: #2063108